pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agent-sentiment-analysis"
version = "0.1.0"
description = "Multi-agent system for sentiment analysis using CrewAI"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "crewai",
    "praw",
    "textblob",
    "python-dotenv",
]
requires-python = ">=3.8"
readme = "README.md"

[project.scripts]
run = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]